refactor(elasticsearch-exporter): extract ElasticClient interface for mocking
